(#6814) - Add support to store cookies on Node.js
On Node.js, PouchDB does not hold a cookie jar for cookies received
from server during authentication. It uses the request Node.js module
to make HTTP requests but it does not configure it correctly to hold
a cookie jar.

This adds the minimal required configuration for the request module.